* 代码整理

+++ agileway-ssh/agileway-sshclient/src/main/java/com/jn/agileway/ssh/client/utils/Buffer.java
@@ -228,7 +228,7 @@ public class Buffer<T extends Buffer<T>> {
      * @return this
      */
     public T putBytes(byte[] b, int off, int len) {
-        return putUInt32(len - off).putRawBytes(b, off, len);
+        return putUInt32((long)len - off).putRawBytes(b, off, len);
     }
 
     public byte[] remainingRawBytes() {
